Finish adding move
This adds move for Shape, Oval, and Button. Shoes-level implementation
and an SWT gui implementation. Oval is straghtforward.  Shape is
somewhat complicated because it moves all elements of the shape. I
believe this is a correct, but not optimized implementation.

Button is also complicated because Buttons can be laid out in slots.
The Button is removed from the slot's layout, so the remaining elements
can reflow. The Button's gui element is disposed, and a new gui element
is created and added to a new Swt::Composite layer so it is out of the
layout flow. This only happens on the first move. Once the Button is in
a container without a layout, it continues to use the same gui element.

* Add NO_BACKGROUND so layers are transparent. This applies to
  all flow objects
* Adjust framerate to approach Red Shoes. It's still not a perfect
  match, but it's closer.
* Add new working sample: simple-move.rb